fixed path variable

+++ kernel/proc.h
@@ -106,19 +106,19 @@ struct proc {
   struct inode *cwd;           // Current directory
   char name[16];               // Process name (debugging)
 
-  uint64 ctime;                // Process creation time
-  uint64 ttime;                // Process termination time
-  uint64 stime;                // The total time the process spent in the SLEEPING mode
-  uint64 retime;               // The total time the process spent in the RUNNABLE mode
-  uint64 rutime;               // The total time the process spent in the RUNNING mode
-  uint64 bursttime;            // Approximate estimated burst time
+  int ctime;                  // Process creation time
+  int ttime;                  // Process termination time
+  int stime;                  // The total time the process spent in the SLEEPING mode
+  int retime;                 // The total time the process spent in the RUNNABLE mode
+  int rutime;                 // The total time the process spent in the RUNNING mode
+  float bursttime;            // Approximate estimated burst time
 };
 
 struct perf {
-  uint64 ctime;                // Process creation time
-  uint64 ttime;                // Process termination time
-  uint64 stime;                // The total time the process spent in the SLEEPING mode
-  uint64 retime;               // The total time the process spent in the RUNNABLE mode
-  uint64 rutime;               // The total time the process spent in the RUNNING mode
-  uint64 bursttime;            // Approximate estimated burst time
+  int ctime;                  // Process creation time
+  int ttime;                  // Process termination time
+  int stime;                  // The total time the process spent in the SLEEPING mode
+  int retime;                 // The total time the process spent in the RUNNABLE mode
+  int rutime;                 // The total time the process spent in the RUNNING mode
+  float bursttime;            // Approximate estimated burst time
 };
